@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -28,15 +28,16 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  google_fonts:
+  google_fonts: ^2.0.0
   # handle screen mobile/tablet/desktop
-  responsive_builder: ^0.3.0
-  get_it: ^6.0.0
-  flutter_riverpod: ^0.13.1
+  responsive_builder: ^0.4.1
+  get_it: ^6.1.1
+  flutter_riverpod: ^0.14.0+2
   intl: ^0.17.0
   dio: ^4.0.0
   tuple: ^2.0.0
-
+  # store user login status
+  # shared_preferences: ^2.0.5
 
 dev_dependencies:
   flutter_test:
