@@ -1,7 +1,7 @@
 class UserModel {
-  int id;
-  String name;
-  String email;
+  int? id;
+  String? name;
+  String? email;
 
   UserModel({this.id, this.name, this.email});
 
