@@ -1,9 +1,9 @@
 class RoutingData {
   final String route;
-  final Map<String, String> _queryParameters;
+  final Map<String, String>? _queryParameters;
 
-  RoutingData({this.route, Map<String, String> queryParameters})
+  RoutingData({required this.route, Map<String, String>? queryParameters})
       : _queryParameters = queryParameters;
 
-  operator [](String key) => _queryParameters[key];
+  operator [](String key) => _queryParameters![key];
 }
