@@ -1,15 +1,16 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:maliya/models/user_model.dart';
 
-final userProvider = StateNotifierProvider<CurrentUser>((ref) {
-  return CurrentUser(null);
+final userProvider = StateNotifierProvider<CurrentUser, UserModel>((ref) {
+  return CurrentUser(new UserModel(id: -1, name: null, email: null));
 });
 
 class CurrentUser extends StateNotifier<UserModel> {
-  UserModel _userModel;
+  UserModel? _userModel;
 
-  CurrentUser(UserModel _userModel)
-      : super(_userModel ?? new UserModel(id: -1, name: null, email: null));
+  // CurrentUser(UserModel _userModel)
+  //     : super(_userModel ?? new UserModel(id: -1, name: null, email: null));
+  CurrentUser(UserModel? _userModel) : super(_userModel!);
 
   // UserModel get userModel => _userModel;
 
@@ -19,6 +20,6 @@ class CurrentUser extends StateNotifier<UserModel> {
   }
 
   void clearUser() {
-    state = null;
+    // state = null;
   }
 }
