@@ -2,7 +2,7 @@ const String RootRoutes = '/';
 const String HomeRoutes = '/home';
 const String EpisodeRoutes = '/episodes';
 const String AboutRoutes = '/about';
-const String MuaRoutes = '/mua';
+const String MalisRoutes = '/malis';
 const String EpisodeDetailRoutes = '/episode';
 const String UnknownRoutes = '/404';
 
